Test fixes after dicussion
