+++ be/__tests__/config/globalSetup.ts
@@ -1,9 +1,8 @@
 import { DatabaseHandler } from '../../src/db/index.js';
 import { HttpServer } from '../../src/server/index.js';
 import { EventEmitter, sql } from '../../src/types/index.js';
-import { ERR_CODES } from '../../src/utils/index.js';
 
-import { cleanupDatabase, mockLogger } from './utils.js';
+import { cleanupDatabase, getTestEnv, mockLogger } from './utils.js';
 
 /**********************************************************************************/
 
@@ -76,61 +75,4 @@ export async function setup({ provide }: Provide) {
     await cleanupDatabase(db);
     server.close();
   };
-}
-
-/**********************************************************************************/
-
-function getTestEnv() {
-  const mode = process.env.NODE_ENV;
-  checkRuntimeEnv(mode);
-  checkEnvVariables();
-
-  return {
-    mode: process.env.NODE_ENV as 'test',
-    server: {
-      base: 'http://localhost',
-      port: process.env.TEST_SERVER_PORT!,
-      apiRoute: process.env.API_ROUTE!,
-      healthCheck: {
-        route: process.env.HEALTH_CHECK_ROUTE!,
-        allowedHosts: new Set(process.env.ALLOWED_HOSTS!.split(','))
-      }
-    },
-    db: process.env.DB_TEST_URI!
-  };
-}
-
-function checkRuntimeEnv(mode?: string | undefined): mode is 'test' {
-  if (mode && mode === 'test') {
-    return true;
-  }
-
-  console.error(
-    `Missing or invalid 'NODE_ENV' env value, should never happen.` +
-      ' Unresolvable, exiting...'
-  );
-
-  process.exit(ERR_CODES.EXIT_NO_RESTART);
-}
-
-function checkEnvVariables() {
-  let missingValues = '';
-  [
-    'TEST_SERVER_PORT',
-    'API_ROUTE',
-    'HEALTH_CHECK_ROUTE',
-    'ALLOWED_HOSTS',
-    'DB_TEST_URI'
-  ].forEach((val) => {
-    if (!process.env[val]) {
-      missingValues += `* Missing ${val} environment variable\n`;
-    }
-  });
-  if (missingValues) {
-    console.error(
-      `\nMissing the following environment vars:\n${missingValues}`
-    );
-
-    process.exit(ERR_CODES.EXIT_NO_RESTART);
-  }
 }
